backend/py_analytics/requirements.txt
Django==5.1.2
djangorestframework==3.14.0
gunicorn==21.2.0
